pyproject.toml
[tool.poetry]
name = "srsly"
version = "0.1.0"
description = "srs.ly backend"
authors = ["Jethro Kuan <jethrokuan95@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
bottle = "^0.12.19"
requests = "^2.25.1"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
app = "srsly.app:main"
